refactor tabbedColumns and multiColumns to use the same code, this will avoid futur bugs
